[#9] Refactor: 폴더 구조 컨벤션 적용
