+++ app/page.tsx
@@ -1,0 +1,7 @@
+import * as styles from "@/app/page.css";
+
+const Home = () => {
+  return <div className={styles.container}>homee</div>;
+};
+
+export default Home;
